miscellaneous/embeddings/vector-embeddings-demos/requirements.txt
matplotlib
numpy
gensim
ipykernel
ipywidgets
scipy==1.12
scikit-learn
openai>=1.109.1
azure-identity
python-dotenv
pandas
hnswlib
